requirements.txt
pytorch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
torcheval
opencv-python
albumentations
tqdm
tensorboard